upgrade: reduce eth_calls
