+++ package.json
@@ -28,7 +28,7 @@
   },
   "devDependencies": {
     "@types/node": "18.7.18",
-    "@types/react": "^18.0.25",
+    "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.2",
     "eslint": "8.23.1",
     "eslint-config-next": "12.3.0",
